crewai-service/requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.20
langchain-google-genai==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.0
requests==2.31.0
